+++ src/op/arc_weight.rs
@@ -27,9 +27,16 @@
 extern crate alloc;
 
 use {
-    alloc::collections::BTreeMap,
+    crate::op::HasArc,
+    alloc::collections::{
+        BTreeMap,
+        BTreeSet,
+    },
     core::hash::BuildHasher,
-    std::collections::HashMap,
+    std::collections::{
+        HashMap,
+        HashSet,
+    },
 };
 
 /// Get the weight of an arc.
@@ -90,7 +97,7 @@ pub trait ArcWeight<W> {
     fn arc_weight(&self, s: usize, t: usize) -> Option<&W>;
 }
 
-macro_rules! impl_usize {
+macro_rules! impl_weighted_usize {
     () => {
         fn arc_weight(&self, s: usize, t: usize) -> Option<&W> {
             self.get(s).and_then(|m| m.get(&t))
@@ -98,7 +105,7 @@ macro_rules! impl_usize {
     };
 }
 
-macro_rules! impl_ref_usize {
+macro_rules! impl_weighted_ref_usize {
     () => {
         fn arc_weight(&self, s: usize, t: usize) -> Option<&W> {
             self.get(&s).and_then(|m| m.get(&t))
@@ -106,48 +113,111 @@ macro_rules! impl_ref_usize {
     };
 }
 
+macro_rules! impl_unweighted {
+    () => {
+        fn arc_weight(&self, s: usize, t: usize) -> Option<&usize> {
+            self.has_arc(s, t).then_some(&1)
+        }
+    };
+}
+
+impl ArcWeight<usize> for Vec<BTreeSet<usize>> {
+    impl_unweighted!();
+}
+
+impl<H> ArcWeight<usize> for Vec<HashSet<usize, H>>
+where
+    H: BuildHasher,
+{
+    impl_unweighted!();
+}
+
+impl ArcWeight<usize> for [BTreeSet<usize>] {
+    impl_unweighted!();
+}
+
+impl<H> ArcWeight<usize> for [HashSet<usize, H>]
+where
+    H: BuildHasher,
+{
+    impl_unweighted!();
+}
+
+impl<const V: usize> ArcWeight<usize> for [BTreeSet<usize>; V] {
+    impl_unweighted!();
+}
+
+impl<const V: usize, H> ArcWeight<usize> for [HashSet<usize, H>; V]
+where
+    H: BuildHasher,
+{
+    impl_unweighted!();
+}
+
+impl ArcWeight<usize> for BTreeMap<usize, BTreeSet<usize>> {
+    impl_unweighted!();
+}
+
+impl<H> ArcWeight<usize> for HashMap<usize, HashSet<usize, H>, H>
+where
+    H: BuildHasher,
+{
+    impl_unweighted!();
+}
+
+impl ArcWeight<usize> for BTreeSet<(usize, usize)> {
+    impl_unweighted!();
+}
+
+impl<H> ArcWeight<usize> for HashSet<(usize, usize), H>
+where
+    H: BuildHasher,
+{
+    impl_unweighted!();
+}
+
 impl<W> ArcWeight<W> for Vec<BTreeMap<usize, W>> {
-    impl_usize!();
+    impl_weighted_usize!();
 }
 
 impl<W, H> ArcWeight<W> for Vec<HashMap<usize, W, H>>
 where
     H: BuildHasher,
 {
-    impl_usize!();
+    impl_weighted_usize!();
 }
 
 impl<W> ArcWeight<W> for [BTreeMap<usize, W>] {
-    impl_usize!();
+    impl_weighted_usize!();
 }
 
 impl<W, H> ArcWeight<W> for [HashMap<usize, W, H>]
 where
     H: BuildHasher,
 {
-    impl_usize!();
+    impl_weighted_usize!();
 }
 
 impl<const V: usize, W> ArcWeight<W> for [BTreeMap<usize, W>; V] {
-    impl_usize!();
+    impl_weighted_usize!();
 }
 
 impl<const V: usize, W, H> ArcWeight<W> for [HashMap<usize, W, H>; V]
 where
     H: BuildHasher,
 {
-    impl_usize!();
+    impl_weighted_usize!();
 }
 
 impl<W> ArcWeight<W> for BTreeMap<usize, BTreeMap<usize, W>> {
-    impl_ref_usize!();
+    impl_weighted_ref_usize!();
 }
 
 impl<W, H> ArcWeight<W> for HashMap<usize, HashMap<usize, W, H>, H>
 where
     H: BuildHasher,
 {
-    impl_ref_usize!();
+    impl_weighted_ref_usize!();
 }
 
 #[cfg(test)]
@@ -159,11 +229,37 @@ mod tests {
                 Empty,
                 EmptyConst,
             },
-            op::AddWeightedArc,
+            op::{
+                AddArc,
+                AddWeightedArc,
+            },
         },
     };
 
-    macro_rules! test_arc_weight {
+    macro_rules! test_unweighted {
+        ($digraph:expr) => {
+            $digraph.add_arc(0, 1);
+            $digraph.add_arc(0, 2);
+            $digraph.add_arc(1, 0);
+            $digraph.add_arc(2, 0);
+            $digraph.add_arc(2, 1);
+
+            assert_eq!($digraph.arc_weight(0, 0), None);
+            assert_eq!($digraph.arc_weight(0, 1), Some(&1));
+            assert_eq!($digraph.arc_weight(0, 2), Some(&1));
+            assert_eq!($digraph.arc_weight(0, 3), None);
+            assert_eq!($digraph.arc_weight(1, 0), Some(&1));
+            assert_eq!($digraph.arc_weight(1, 1), None);
+            assert_eq!($digraph.arc_weight(1, 2), None);
+            assert_eq!($digraph.arc_weight(1, 3), None);
+            assert_eq!($digraph.arc_weight(2, 0), Some(&1));
+            assert_eq!($digraph.arc_weight(2, 1), Some(&1));
+            assert_eq!($digraph.arc_weight(2, 2), None);
+            assert_eq!($digraph.arc_weight(2, 3), None);
+        };
+    }
+
+    macro_rules! test_weighted {
         ($digraph:expr) => {
             $digraph.add_weighted_arc(0, 1, 2);
             $digraph.add_weighted_arc(0, 2, 3);
@@ -171,67 +267,144 @@ mod tests {
             $digraph.add_weighted_arc(2, 0, 7);
             $digraph.add_weighted_arc(2, 1, 8);
 
+            assert_eq!($digraph.arc_weight(0, 0), None);
             assert_eq!($digraph.arc_weight(0, 1), Some(&2));
             assert_eq!($digraph.arc_weight(0, 2), Some(&3));
+            assert_eq!($digraph.arc_weight(0, 3), None);
             assert_eq!($digraph.arc_weight(1, 0), Some(&4));
+            assert_eq!($digraph.arc_weight(1, 1), None);
+            assert_eq!($digraph.arc_weight(1, 2), None);
+            assert_eq!($digraph.arc_weight(1, 3), None);
             assert_eq!($digraph.arc_weight(2, 0), Some(&7));
             assert_eq!($digraph.arc_weight(2, 1), Some(&8));
+            assert_eq!($digraph.arc_weight(2, 2), None);
+            assert_eq!($digraph.arc_weight(2, 3), None);
         };
+    }
+
+    #[test]
+    fn vec_btree_set() {
+        let mut digraph = Vec::<BTreeSet<usize>>::empty(3);
+
+        test_unweighted!(digraph);
+    }
+
+    #[test]
+    fn vec_hash_set() {
+        let mut digraph = Vec::<HashSet<usize>>::empty(3);
+
+        test_unweighted!(digraph);
+    }
+
+    #[test]
+    fn slice_btree_set() {
+        let mut digraph = Vec::<BTreeSet<usize>>::empty(3);
+
+        test_unweighted!(digraph.as_mut_slice());
+    }
+
+    #[test]
+    fn slice_hash_set() {
+        let mut digraph = Vec::<HashSet<usize>>::empty(3);
+
+        test_unweighted!(digraph.as_mut_slice());
+    }
+
+    #[test]
+    fn arr_btree_set() {
+        let mut digraph = <[BTreeSet<usize>; 3]>::empty();
+
+        test_unweighted!(digraph);
+    }
+
+    #[test]
+    fn arr_hash_set() {
+        let mut digraph = <[HashSet<usize>; 3]>::empty();
+
+        test_unweighted!(digraph);
+    }
+
+    #[test]
+    fn btree_map_btree_set() {
+        let mut digraph = BTreeMap::<usize, BTreeSet<usize>>::empty(3);
+
+        test_unweighted!(digraph);
+    }
+
+    #[test]
+    fn hash_map_hash_set() {
+        let mut digraph = HashMap::<usize, HashSet<usize>>::empty(3);
+
+        test_unweighted!(digraph);
     }
 
     #[test]
     fn vec_btree_map() {
         let mut digraph = Vec::<BTreeMap<usize, i32>>::empty(3);
 
-        test_arc_weight!(digraph);
+        test_weighted!(digraph);
     }
 
     #[test]
     fn vec_hash_map() {
         let mut digraph = Vec::<HashMap<usize, i32>>::empty(3);
 
-        test_arc_weight!(digraph);
+        test_weighted!(digraph);
     }
 
     #[test]
     fn slice_btree_map() {
         let mut digraph = Vec::<BTreeMap<usize, i32>>::empty(3);
 
-        test_arc_weight!(digraph.as_mut_slice());
+        test_weighted!(digraph.as_mut_slice());
     }
 
     #[test]
     fn slice_hash_map() {
         let mut digraph = Vec::<HashMap<usize, i32>>::empty(3);
 
-        test_arc_weight!(digraph.as_mut_slice());
+        test_weighted!(digraph.as_mut_slice());
     }
 
     #[test]
     fn arr_btree_map() {
         let mut digraph = <[BTreeMap<usize, i32>; 3]>::empty();
 
-        test_arc_weight!(digraph);
+        test_weighted!(digraph);
     }
 
     #[test]
     fn arr_hash_map() {
         let mut digraph = <[HashMap<usize, i32>; 3]>::empty();
 
-        test_arc_weight!(digraph);
+        test_weighted!(digraph);
     }
 
     #[test]
     fn btree_map_btree_map() {
         let mut digraph = BTreeMap::<usize, BTreeMap<usize, i32>>::empty(3);
 
-        test_arc_weight!(digraph);
+        test_weighted!(digraph);
     }
 
     #[test]
     fn hash_map_hash_map() {
         let mut digraph = HashMap::<usize, HashMap<usize, i32>>::empty(3);
 
-        test_arc_weight!(digraph);
+        test_weighted!(digraph);
+    }
+
+    #[test]
+    fn btree_set_tuple() {
+        let mut digraph = BTreeSet::<(usize, usize)>::empty(3);
+
+        test_unweighted!(digraph);
+    }
+
+    #[test]
+    fn hash_set_tuple() {
+        let mut digraph = HashSet::<(usize, usize)>::empty(3);
+
+        test_unweighted!(digraph);
     }
 }